Split share_item_service into S3 specific and generic
